+++ keep-ui/components/navbar/CustomPresetAlertLinks.tsx
@@ -18,7 +18,7 @@ import {
 } from "@dnd-kit/core";
 import { SortableContext, useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
-import { Preset } from "app/alerts/models";
+import { Preset } from "@/app/(keep)/alerts/models";
 import { AiOutlineSound } from "react-icons/ai";
 // Using dynamic import to avoid hydration issues with react-player
 import dynamic from "next/dynamic";
