--- keep-ui/types/next-auth.d.ts
@@ -1,20 +1,0 @@
-import NextAuth from "next-auth";
-
-declare module "next-auth" {
-  /**
-   * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
-   */
-  interface Session {
-    user: User;
-    accessToken?: string;
-    tenantId?: string;
-    userRole?: string;
-  }
-  interface User {
-    name: string;
-    email: string;
-    accessToken: string;
-    tenantId?: string;
-    role?: string;
-  }
-}
